+++ examples/ai-core/src/generate-text/google-vertex-anthropic-tool-call.ts
@@ -1,0 +1,58 @@
+import 'dotenv/config';
+import { vertexAnthropic } from '@ai-sdk/google-vertex/anthropic';
+import { generateText, tool } from 'ai';
+import { z } from 'zod';
+import { weatherTool } from '../tools/weather-tool';
+
+async function main() {
+  const result = await generateText({
+    model: vertexAnthropic('claude-3-5-sonnet-v2@20241022'),
+    maxTokens: 512,
+    tools: {
+      weather: weatherTool,
+      cityAttractions: tool({
+        parameters: z.object({ city: z.string() }),
+      }),
+    },
+    prompt:
+      'What is the weather in San Francisco and what attractions should I visit?',
+  });
+
+  // typed tool calls:
+  for (const toolCall of result.toolCalls) {
+    switch (toolCall.toolName) {
+      case 'cityAttractions': {
+        toolCall.args.city; // string
+        break;
+      }
+
+      case 'weather': {
+        toolCall.args.location; // string
+        break;
+      }
+    }
+  }
+
+  // typed tool results for tools with execute method:
+  for (const toolResult of result.toolResults) {
+    switch (toolResult.toolName) {
+      // NOT AVAILABLE (NO EXECUTE METHOD)
+      // case 'cityAttractions': {
+      //   toolResult.args.city; // string
+      //   toolResult.result;
+      //   break;
+      // }
+
+      case 'weather': {
+        toolResult.args.location; // string
+        toolResult.result.location; // string
+        toolResult.result.temperature; // number
+        break;
+      }
+    }
+  }
+
+  console.log(JSON.stringify(result, null, 2));
+}
+
+main().catch(console.error);