+++ examples/ai-core/src/stream-text/google-vertex-anthropic-pdf.ts
@@ -1,0 +1,32 @@
+import 'dotenv/config';
+import { vertexAnthropic } from '@ai-sdk/google-vertex/anthropic';
+import { streamText } from 'ai';
+import fs from 'node:fs';
+
+async function main() {
+  const result = streamText({
+    model: vertexAnthropic('claude-3-5-sonnet-v2@20241022'),
+    messages: [
+      {
+        role: 'user',
+        content: [
+          {
+            type: 'text',
+            text: 'What is an embedding model according to this document?',
+          },
+          {
+            type: 'file',
+            data: fs.readFileSync('./data/ai.pdf'),
+            mimeType: 'application/pdf',
+          },
+        ],
+      },
+    ],
+  });
+
+  for await (const textPart of result.textStream) {
+    process.stdout.write(textPart);
+  }
+}
+
+main().catch(console.error);