+++ packages/openai-compatible/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/openai-compatible
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [0984f0b]
+  - @ai-sdk/provider-utils@2.0.3
+
 ## 0.0.3
 
 ### Patch Changes