+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/provider-utils
 
+## 2.0.3
+
+### Patch Changes
+
+- 0984f0b: feat (provider-utils): Add resolvable type and utility routine.
+
 ## 2.0.2
 
 ### Patch Changes