+++ packages/swarm/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/swarm
 
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies [9f32213]
+  - ai@4.0.13
+
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies [5167bec]
+- Updated dependencies [0984f0b]
+  - ai@4.0.12
+
 ## 0.0.20
 
 ### Patch Changes